refactor cli commands
