+++ packages/jaspr_pad/templates/jaspr_export/pubspec.yaml
@@ -14,3 +14,4 @@ dev_dependencies:
   build_runner: ^2.1.4
   build_web_compilers: ^3.2.1
   lints: ^1.0.0
+  webdev: 2.7.7
